+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add validation for empty git diff before generating commit message
+- Add unit tests for get_public_ip
 
 ## [0.6.6] - 2026-08-03